script/python/requirements.txt
mysql-connector-python==8.0.30
numpy==1.23.2
pandas==1.4.3
protobuf==3.20.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
six==1.16.0
